+++ survey_file_upload/__manifest__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 {
     'name': 'Survey File Upload',
-    'version': '19.0.1.1.0',
+    'version': '19.0.1.1.1',
     'category': 'Marketing/Surveys',
     'summary': 'Adds a "File Upload" question type to Survey, with AJAX (non-blocking) upload',
     'description': """